+++ NEWS.md
@@ -8,6 +8,7 @@
 * Screen and markdown wrap respect word boundaries.
 * `to_screen` and `to_md` gain a `min_width` argument; `to_md` gains a logical `header` argument; `to_screen` gains
   a `compact` argument replacing `blank = NULL`.
+* On screen colour and bold support, if the `crayon` package is installed.
 
 ## Breaking changes